+++ muk_web_preview_attachment/doc/changelog.rst
@@ -1,3 +1,9 @@
+`2.0.0`
+-------
+
+- Migrated to Python 3
+- Support for FieldMany2ManyBinaryMultiFiles Widget
+
 `1.0.0`
 -------